requirements.txt
asgiref==3.9.1
Django==5.2.5
sqlparse==0.5.3
tzdata==2025.2

djangorestframework>=3.15.0
django-filter>=24.2
djangorestframework-simplejwt>=5.3.0
python-decouple>=3.8
dj-database-url>=2.1.0
Pillow>=10.2.0
django-cloudinary-storage>=0.3.0
channels>=4.0.0
channels-redis>=4.2.0
redis>=5.0.1
celery>=5.3.6.
# Deployment
gunicorn>=21.2.0
whitenoise>=6.6.0  # Serve static files in production
# Testing & Dev Tools
pytest>=8.2.0
pytest-django>=4.8.0
flake8>=7.0.0
black>=24.4.0
# Optional: Containerization
# docker>=7.0.0
